requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
httpx
cryptography
pypdf
python-docx
Pillow
waitress